Add enricher
