+++ hypertrace-trace-enricher/enriched-span-constants/src/main/java/org/hypertrace/traceenricher/enrichedspan/constants/EnrichedSpanConstants.java
@@ -9,6 +9,8 @@ public class EnrichedSpanConstants {
   private static final Logger LOGGER = LoggerFactory.getLogger(EnrichedSpanConstants.class);
 
   public static final String SPACE_IDS_ATTRIBUTE = "SPACE_IDS";
+  public static final String API_EXIT_CALLS_ATTRIBUTE = "API_EXIT_CALLS";
+  public static final String API_EXIT_SERVICES_ATTRIBUTE = "API_EXIT_SERVICES";
 
   /**
    * Returns the constant value for the given Enum.
