+++ hypertrace-trace-enricher/hypertrace-trace-enricher/src/test/resources/configs/hypertrace-trace-enricher/application.conf
@@ -19,7 +19,7 @@ kafka.streams.config = {
 }
 
 enricher {
-  names = ["SpanTypeAttributeEnricher", "ApiStatusEnricher", "EndpointEnricher", "TransactionNameEnricher", "ApiBoundaryTypeAttributeEnricher", "ErrorsAndExceptionsEnricher", "BackendEntityEnricher", "HttpAttributeEnricher", "DefaultServiceEntityEnricher", "UserAgentSpanEnricher"]
+  names = ["SpanTypeAttributeEnricher", "ApiStatusEnricher", "EndpointEnricher", "TransactionNameEnricher", "ApiBoundaryTypeAttributeEnricher", "ErrorsAndExceptionsEnricher", "BackendEntityEnricher", "HttpAttributeEnricher", "DefaultServiceEntityEnricher", "UserAgentSpanEnricher", "ExitCallsEnricher"]
 
   clients = {
       entity.service.config = {
@@ -80,5 +80,9 @@ enricher {
   EndpointEnricher {
     class = "org.hypertrace.traceenricher.enrichment.enrichers.endpoint.EndpointEnricher"
     dependencies = ["DefaultServiceEntityEnricher", "ApiBoundaryTypeAttributeEnricher"]
+  }
+
+  ExitCallsEnricher {
+    class = "org.hypertrace.traceenricher.enrichment.enrichers.ExitCallsEnricher"
   }
 }
