+++ hypertrace-view-generator/hypertrace-view-generator-api/src/main/avro/SpanEventView.avdl
@@ -100,6 +100,6 @@ protocol SpanEventViewProtocol {
     int api_exit_calls = 0;
 
     // map of service -> exit call count for api
-    map<int> api_exit_services = {};
+    map<string> api_exit_services = {};
   }
 }
