+++ hypertrace-view-generator/hypertrace-view-generator/src/main/java/org/hypertrace/viewgenerator/generators/SpanEventViewGenerator.java
@@ -2,33 +2,26 @@ package org.hypertrace.viewgenerator.generators;
 
 import static org.hypertrace.core.datamodel.shared.SpanAttributeUtils.getStringAttribute;
 
-import com.google.common.collect.Maps;
 import java.nio.ByteBuffer;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
-import java.util.concurrent.atomic.AtomicInteger;
 import java.util.stream.Collectors;
 import org.apache.avro.Schema;
-import org.hypertrace.core.datamodel.ApiNodeEventEdge;
 import org.hypertrace.core.datamodel.Entity;
 import org.hypertrace.core.datamodel.Event;
 import org.hypertrace.core.datamodel.MetricValue;
 import org.hypertrace.core.datamodel.StructuredTrace;
 import org.hypertrace.core.datamodel.eventfields.http.Http;
 import org.hypertrace.core.datamodel.eventfields.http.Request;
-import org.hypertrace.core.datamodel.shared.ApiNode;
 import org.hypertrace.core.datamodel.shared.SpanAttributeUtils;
-import org.hypertrace.entity.constants.v1.BackendAttribute;
-import org.hypertrace.entity.service.constants.EntityConstants;
 import org.hypertrace.traceenricher.enrichedspan.constants.EnrichedSpanConstants;
 import org.hypertrace.traceenricher.enrichedspan.constants.utils.EnrichedSpanUtils;
 import org.hypertrace.traceenricher.enrichedspan.constants.v1.Api;
 import org.hypertrace.traceenricher.enrichedspan.constants.v1.CommonAttribute;
 import org.hypertrace.traceenricher.enrichedspan.constants.v1.ErrorMetrics;
 import org.hypertrace.traceenricher.enrichedspan.constants.v1.Protocol;
-import org.hypertrace.traceenricher.trace.util.ApiTraceGraph;
 import org.hypertrace.viewgenerator.api.SpanEventView;
 
 public class SpanEventViewGenerator extends BaseViewGenerator<SpanEventView> {
@@ -68,8 +61,6 @@ public class SpanEventViewGenerator extends BaseViewGenerator<SpanEventView> {
             parentToChildrenEventIds,
             eventMap);
 
-    Map<ByteBuffer, ApiExitCallInfo> eventToApiExitCall = computeApiExitCallCount(structuredTrace);
-
     return structuredTrace.getEventList().stream()
         .map(
             event ->
@@ -78,8 +69,7 @@ public class SpanEventViewGenerator extends BaseViewGenerator<SpanEventView> {
                         structuredTrace.getTraceId(),
                         eventMap,
                         childToParentEventIds,
-                        exitSpanToCalleeApiEntrySpanMap,
-                        eventToApiExitCall)
+                        exitSpanToCalleeApiEntrySpanMap)
                     .build())
         .collect(Collectors.toList());
   }
@@ -134,111 +124,12 @@ public class SpanEventViewGenerator extends BaseViewGenerator<SpanEventView> {
         .orElse(null);
   }
 
-  /**
-   * An api_node represents an api call in the trace It consists of api_entry span and multiple
-   * api_exit and internal spans
-   *
-   * <p>This method computes the count of exit calls for any given api (identified by
-   * api_entry_span) This count is a composition of 2 things 1. link between api_exit_span in
-   * api_node to api_entry_span (child of api_exit_span) in another api_node 2. exit calls to
-   * backend from api_exit_span in api_node
-   */
-  Map<ByteBuffer, ApiExitCallInfo> computeApiExitCallCount(StructuredTrace trace) {
-    ApiTraceGraph apiTraceGraph = ViewGeneratorState.getApiTraceGraph(trace);
-    // event -> api exit call count for the corresponding api_node
-    Map<ByteBuffer, ApiExitCallInfo> eventToExitInfo = Maps.newHashMap();
-
-    for (ApiNode<Event> apiNode : apiTraceGraph.getApiNodeList()) {
-      List<Event> backendExitEvents =
-          apiTraceGraph.getExitBoundaryEventsWithNoOutboundEdgeForApiNode(apiNode);
-      List<ApiNodeEventEdge> edges = apiTraceGraph.getOutboundEdgesForApiNode(apiNode);
-      int totalExitCallCount = backendExitEvents.size() + edges.size();
-      ApiExitCallInfo apiExitCallInfo = new ApiExitCallInfo().withExitCallCount(totalExitCallCount);
-      edges.forEach(edge -> apiExitCallInfo.handleApiNodeEdge(trace, edge));
-      backendExitEvents.forEach(apiExitCallInfo::handleBackend);
-      apiNode.getEvents().forEach(e -> eventToExitInfo.put(e.getEventId(), apiExitCallInfo));
-    }
-    return eventToExitInfo;
-  }
-
-  static class ApiExitCallInfo {
-
-    private static final String UNKNOWN_SERVICE = "unknown-service";
-    private static final String UNKNOWN_BACKEND = "unknown-backend";
-
-    private final Map<String, String> calleeIdToName;
-    private final Map<String, AtomicInteger> calleeIdToCallCount;
-
-    private int exitCallCount;
-    private int unknownServiceExits;
-    private int unknownBackendExits;
-
-    public ApiExitCallInfo() {
-      unknownServiceExits = 0;
-      unknownBackendExits = 0;
-      this.calleeIdToName = Maps.newHashMap();
-      this.calleeIdToCallCount = Maps.newHashMap();
-    }
-
-    public ApiExitCallInfo withExitCallCount(int exitCallCount) {
-      this.exitCallCount = exitCallCount;
-      return this;
-    }
-
-    void handleApiNodeEdge(StructuredTrace trace, ApiNodeEventEdge edge) {
-      Event event = trace.getEventList().get(edge.getTgtEventIndex());
-      if (EnrichedSpanUtils.containsServiceId(event)) {
-        calleeIdToName.put(
-            EnrichedSpanUtils.getServiceId(event), EnrichedSpanUtils.getServiceName(event));
-        calleeIdToCallCount
-            .computeIfAbsent(EnrichedSpanUtils.getServiceId(event), v -> new AtomicInteger(0))
-            .incrementAndGet();
-      } else {
-        unknownServiceExits++;
-      }
-    }
-
-    void handleBackend(Event exitEvent) {
-      if (exitEvent
-          .getEnrichedAttributes()
-          .getAttributeMap()
-          .containsKey(EntityConstants.getValue(BackendAttribute.BACKEND_ATTRIBUTE_ID))) {
-        calleeIdToName.put(
-            EnrichedSpanUtils.getBackendId(exitEvent), EnrichedSpanUtils.getBackendName(exitEvent));
-        calleeIdToCallCount
-            .computeIfAbsent(EnrichedSpanUtils.getBackendId(exitEvent), v -> new AtomicInteger(0))
-            .incrementAndGet();
-      } else {
-        unknownBackendExits++;
-      }
-    }
-
-    int getExitCallCount() {
-      return exitCallCount;
-    }
-
-    Map<String, Integer> getCalleeNameToExitCalls() {
-      Map<String, Integer> calleeNameToExitCalls =
-          calleeIdToCallCount.entrySet().stream()
-              .collect(
-                  Collectors.toMap(k -> calleeIdToName.get(k.getKey()), v -> v.getValue().get()));
-      if (unknownServiceExits > 0) {
-        calleeNameToExitCalls.put(UNKNOWN_SERVICE, unknownServiceExits);
-      }
-      if (unknownBackendExits > 0) {
-        calleeNameToExitCalls.put(UNKNOWN_BACKEND, unknownBackendExits);
-      }
-      return calleeNameToExitCalls;
-    }
-  }
-
   private SpanEventView.Builder generateViewBuilder(
       Event event,
       ByteBuffer traceId,
       Map<ByteBuffer, Event> eventMap,
       Map<ByteBuffer, ByteBuffer> childToParentEventIds,
-      Map<ByteBuffer, Event> exitSpanToCalleeApiEntrySpanMap,
-      Map<ByteBuffer, ApiExitCallInfo> eventToApiExitCall) {
+      Map<ByteBuffer, Event> exitSpanToCalleeApiEntrySpanMap) {
 
     SpanEventView.Builder builder = SpanEventView.newBuilder();
 
@@ -254,10 +145,13 @@ public class SpanEventViewGenerator extends BaseViewGenerator<SpanEventView> {
     if (event.getEventId().equals(apiEntrySpanId)) {
       // set this count to 1 only if this span is the head of the Api Trace
       builder.setApiTraceCount(1);
-      builder.setApiExitServices(
-          eventToApiExitCall
-              .getOrDefault(event.getEventId(), new ApiExitCallInfo())
-              .getCalleeNameToExitCalls());
+      if (SpanAttributeUtils.containsAttributeKey(
+          event, EnrichedSpanConstants.API_EXIT_SERVICES_ATTRIBUTE)) {
+        builder.setApiExitServices(
+            SpanAttributeUtils.getAttributeValue(
+                    event, EnrichedSpanConstants.API_EXIT_SERVICES_ATTRIBUTE)
+                .getValueMap());
+      }
     } else {
       builder.setApiTraceCount(0);
     }
@@ -340,9 +234,9 @@ public class SpanEventViewGenerator extends BaseViewGenerator<SpanEventView> {
     builder.setSpaceIds(EnrichedSpanUtils.getSpaceIds(event));
 
     builder.setApiExitCalls(
-        eventToApiExitCall
-            .getOrDefault(event.getEventId(), new ApiExitCallInfo())
-            .getExitCallCount());
+        Integer.parseInt(
+            SpanAttributeUtils.getStringAttributeWithDefault(
+                event, EnrichedSpanConstants.API_EXIT_CALLS_ATTRIBUTE, "0")));
 
     return builder;
   }
